fix(pad): Change return type to isize.

+++ src/utils.rs
@@ -9,6 +9,6 @@ pub fn get_default_display_number() -> Result<u16, env::VarError> {
 ///
 /// See this for more:
 /// https://www.x.org/releases/X11R7.7/doc/xproto/x11protocol.html#Syntactic_Conventions_b
-pub fn pad(e: usize) -> isize {
-    ((4 - (e as isize % 4)) % 4)
+pub fn pad(e: usize) -> usize {
+    ((4 - (e % 4)) % 4)
 }
